+++ qbd-frontend/src/app/components/homepage/component.homepage.ts
@@ -3,7 +3,7 @@ import { app_book_list } from "../book_list/component.book_list";
 import { app_book_filter } from '../filter/component.filter';
 import { BookService } from '../../services/book_service';
 import { Book } from '../../models/book';
-import { tag_type } from '../../tags/tag_type';
+import { tag_type } from '../../models/tag_type';
 
 @Component({
     selector: 'app-startpage',
