+++ qbd-frontend/src/app/models/book.ts
@@ -1,4 +1,4 @@
-import { tag } from "../tags/tag";
+import { tag } from "./tag";
 export interface Book{
     id: string;
     title: string;
